feat: dedicated NuGetClient per NuGet feed
Have the client factory create a NuGetClient instead of an HttpClient. This seems like a better approach.

NOTE: the GitHub feed schema is different from the current NuGet feed schema and does not deserialize properly with the current model. For now only processing the default NuGet feed in the handler.