feat: fix the Semantic Version problem
Some of the NuGet feed items have invalid semantic versions. To keep track of that I added the string version to the class and an IsValid property, that should make it easy to spot in the logs if necessary.
For now the properties are not used, simply added for keeping track of things in case things go wrong.