fix: fix tool package url (#50)
Fix the package URL issue that popped up in the latest dependency update
and build.

There is an issue with a double slash in the tool package URL which
seems to only fail in some circumstances, in this case in GitHub
actions. Fixed it by trimming trailing slashes in the
registrationBaseUrl.
Will add an issue to review since it might make sense to use a different
URL from the service index.

Closes #49